- Change to using floats for coordinates

+++ capi/src/implementation/visibility_graphs/code/datastructures/graph/graph.cpp
@@ -111,8 +111,8 @@ bool Graph::operator!=(const Graph &other) const { return !(*this == other); }
 
 struct AStarHeapElement {
     Coordinate node;
-    double distance_to_source;
-    double heuristic_distance_to_destination;
+    float distance_to_source;
+    float heuristic_distance_to_destination;
 };
 
 std::vector<Coordinate> Graph::shortest_path(const Coordinate &source, const Coordinate &destination) const {
@@ -162,7 +162,7 @@ std::vector<Coordinate> Graph::shortest_path(const Coordinate &source, const Coo
     });
 
     auto prev_coord = std::unordered_map<Coordinate, Coordinate>();
-    auto distances_to_source = std::unordered_map<Coordinate, double>();
+    auto distances_to_source = std::unordered_map<Coordinate, float>();
     while (!pq.empty()) {
         const auto top = pq.top();
         pq.pop();
